+++ src/feature/like/hook/index.js
@@ -1,0 +1,1 @@
+export { default as useLike } from './useLike';
