+++ src/hooks/index.js
@@ -1,3 +1,2 @@
-export { default as useLike } from './useLike';
 export { default as useScrap } from './useScrap';
 export { default as useSearch } from './useSearch';
