+++ src/pages/PostPage/PostPage/PostPage.jsx
@@ -2,6 +2,8 @@ import { useState } from 'react';
 import { useParams, useLocation, useNavigate } from 'react-router-dom';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
+import { getPostContent, deletePost, reportPost, reportUser } from '@/apis';
+
 import { useAuth, useToast } from '@/shared/hook';
 import {
   BackAppBar,
@@ -15,9 +17,9 @@ import { MUTATION_KEY, QUERY_KEY, TOAST, LIKE_TYPE } from '@/shared/constant';
 
 import { useCommentContext } from '@/feature/comment/context';
 import { CommentInput, CommentsSuspense } from '@/feature/comment/component';
+import { useLike } from '@/feature/like/hook';
+import { useScrap } from '@/hooks';
 
-import { getPostContent, deletePost, reportPost, reportUser } from '@/apis';
-import { useLike, useScrap } from '@/hooks';
 import { PostContent } from '@/pages/PostPage';
 import { NotFoundPage } from '@/pages/NotFoundPage';
